Moving more code in testutils
